pyproject.toml
[tool.poetry]
name = "pdf-comparison-worker"
version = "0.1.0"
description = ""
authors = ["Jeremy Herzog <jeremyaherzog@gmail.com>"]
readme = "README.md"
packages = [{include = "worker"}]

[tool.poetry.dependencies]
python = "^3.10"
celery = {extras = ["redis"], version = "^5.4.0"}
redis = "^5.1.1"
pydantic = "^2.9.2"
s3fs = "^2024.9.0"
supabase = "^2.9.0"
pymupdf = "^1.24.11"
fastapi = "^0.115.0"
modal = "^0.64.178"
transformers = "^4.45.2"
python-dotenv = "^1.0.1"
pillow = "^10.4.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
